wip: Resolve comments on PR
