+++ src/Form/grid/StyledContainer/index.tsx
@@ -1,19 +1,14 @@
-import React, {
-  PropsWithChildren,
-  forwardRef,
-  useEffect,
-  useMemo,
-  useRef
-} from 'react';
+import React, { PropsWithChildren, forwardRef } from 'react';
 import {
   useContainerEngine,
   useContainerStyles,
   useFormattedNode,
   useNodeType
 } from './hooks';
+import { FORM_Z_INDEX } from '../../../utils/styles';
 import { getCellStyle } from './styles';
+import { useFixedContainer } from './hooks/useFixedContainer';
 import classNames from 'classnames';
-import { featheryDoc } from '../../../utils/browser';
 
 export type StyledContainerProps = PropsWithChildren & {
   key?: string;
@@ -49,9 +44,6 @@ export const StyledContainer = forwardRef<HTMLDivElement, StyledContainerProps>(
   ) => {
     const { node, rawNode } = useFormattedNode(_node, raw);
     const type = useNodeType(node, rawNode, viewport);
-    const fixedContainerRef = useRef<HTMLDivElement>(null);
-    const nodeRef = useRef(node || rawNode);
-    const getNode = () => (nodeRef.current ? nodeRef.current : node || rawNode);
 
     const { styles, innerStyles, isFixed } = useContainerStyles(
       node,
@@ -62,83 +54,7 @@ export const StyledContainer = forwardRef<HTMLDivElement, StyledContainerProps>(
 
     useContainerEngine(node, rawNode, ref);
 
-    useEffect(() => {
-      nodeRef.current = node || rawNode;
-    }, [node, rawNode]);
-
-    useEffect(() => {
-      let _node = getNode();
-
-      if (_node.uuid || !isFixed) {
-        return; // Disable all fixed behaviour on the editor
-      }
-
-      if (fixedContainerRef.current) {
-        const doc = featheryDoc();
-        const container = doc.querySelector(`div[data-id="${_node.id}"]`);
-
-        if (container) {
-          // If the container is the first child of root, top/left will be 0
-          // If the container is the last child of root, bottom/left will be 0
-          // If the container is none of the above, top/left will be taken from the container
-          // Width will be taken from the container
-
-          const setPositionAndDimensions = () => {
-            if (!fixedContainerRef.current || !container) {
-              return;
-            }
-
-            _node = getNode();
-
-            const { top, left } = container.getBoundingClientRect();
-            const isParentRoot = !_node.parent.parent;
-            const isFirstChild = _node.parent.children[0].id === _node.id;
-            const isLastChild =
-              _node.parent.children[_node.parent.children.length - 1].id ===
-              _node.id;
-
-            if (isParentRoot && isFirstChild) {
-              fixedContainerRef.current.style.top = '0';
-              fixedContainerRef.current.style.left = `${left}px`;
-            }
-
-            if (isParentRoot && isLastChild) {
-              fixedContainerRef.current.style.bottom = '0';
-              fixedContainerRef.current.style.left = `${left}px`;
-            }
-
-            if (!isParentRoot || (!isFirstChild && !isLastChild)) {
-              fixedContainerRef.current.style.top = `${top}px`;
-              fixedContainerRef.current.style.left = `${left}px`;
-            }
-
-            fixedContainerRef.current.style.height = `${container.offsetHeight}px`;
-            fixedContainerRef.current.style.width = `${container.offsetWidth}px`;
-          };
-
-          setPositionAndDimensions();
-
-          let resizeObserver: any = new ResizeObserver(() =>
-            setPositionAndDimensions()
-          );
-
-          let parentResizeObserver: any = new ResizeObserver(() =>
-            setPositionAndDimensions()
-          );
-
-          resizeObserver.observe(container);
-          parentResizeObserver.observe(doc.querySelector('body'));
-
-          return () => {
-            resizeObserver.disconnect();
-            resizeObserver = null;
-
-            parentResizeObserver.disconnect();
-            parentResizeObserver = null;
-          };
-        }
-      }
-    }, [isFixed]);
+    const fixedContainerRef = useFixedContainer(node, rawNode, isFixed);
 
     if (component) {
       const Component = component;
@@ -171,7 +87,7 @@ export const StyledContainer = forwardRef<HTMLDivElement, StyledContainerProps>(
             css={{
               ...styles,
               position: 'fixed',
-              zIndex: 50
+              zIndex: FORM_Z_INDEX + 1
             }}
             ref={fixedContainerRef}
           >
