+++ src/elements/fields/ButtonGroupField.tsx
@@ -1,6 +1,7 @@
 import React, { useMemo } from 'react';
 import { imgMaxSizeStyles, noTextSelectStyles } from '../styles';
 import useBorder from '../components/useBorder';
+import { FORM_Z_INDEX } from '../../utils/styles';
 
 function ButtonGroupField({
   element,
@@ -118,7 +119,7 @@ function ButtonGroupField({
             position: 'absolute',
             bottom: 0,
             opacity: 0,
-            zIndex: -1
+            zIndex: FORM_Z_INDEX - 2
           }}
         />
       </div>
