+++ src/elements/fields/PasswordField.tsx
@@ -5,6 +5,7 @@ import InlineTooltip from '../components/InlineTooltip';
 import { bootstrapStyles } from '../styles';
 import useBorder from '../components/useBorder';
 import BorderlessEyeIcon from '../components/icons/BorderlessEyeIcon';
+import { FORM_Z_INDEX } from '../../utils/styles';
 
 function PasswordField({
   element,
@@ -62,7 +63,7 @@ function PasswordField({
           css={{
             position: 'relative',
             // Position input above the border div
-            zIndex: 1,
+            zIndex: FORM_Z_INDEX,
             height: '100%',
             width: '100%',
             border: 'none',
@@ -101,7 +102,7 @@ function PasswordField({
               right: '8px',
               // We need to subtract half the height of the icon to center it
               top: 'calc(50% - 12px)',
-              zIndex: 1
+              zIndex: FORM_Z_INDEX
             }}
           >
             <BorderlessEyeIcon
