+++ src/elements/fields/TextField/TextAutocomplete.tsx
@@ -1,6 +1,7 @@
 import React, { memo } from 'react';
 
 import { OverlayTrigger } from 'react-bootstrap';
+import { FORM_Z_INDEX } from '../../../utils/styles';
 
 function TextAutocomplete({
   allOptions = [],
@@ -31,7 +32,7 @@ function TextAutocomplete({
         overlay={
           <ul
             css={{
-              zIndex: 1,
+              zIndex: FORM_Z_INDEX,
               listStyleType: 'none',
               padding: 0,
               margin: 0,
